update: introduce pages & components
 fix lint
