+++ src/components/Clients/index.ts
@@ -1,0 +1,4 @@
+import BackupConfig from './BackupConfig.vue';
+import RestoreConfig from './RestoreConfig.vue';
+
+export { BackupConfig, RestoreConfig };
